refactor: привести файловую систему к правилам fsd
